Created Generator and Discriminator Classes among other things
Some of the early work on this project was based on some misunderstandings of how a Convolutional GAN is constructed, and the loss function for them. This is intending to fix those errors. The Discriminator is now set to use a Least Square loss function based off of the dot product of the output vector and the classification vector.